new function random and console

+++ src/spike/control.py
@@ -4,5 +4,5 @@ class wait_for_seconds:
     
     def __init__(self,second):
        # if(settings.ISDEBUG):print (str(second)+' sec wait_for_seconds ended')
-        time.sleep_ms(1000*second)
+        time.sleep_ms(int(1000*second))
       #  if(settings.ISDEBUG):print (str(second)+' sec wait_for_seconds ended')
